pyproject.toml
[project]
name = "wrestling-stats"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
	"duckdb>=1.0.0",
	"playwright>=1.46.0",
	"beautifulsoup4>=4.12.0",
	"pandas>=2.2.0",
	"tqdm>=4.67.1",
	"python-dotenv>=1.0.0",
	"httpx>=0.28.1",
]

[project.optional-dependencies]
optional = []
